+++ src/SelectionToolbar.h
@@ -9,3 +9,4 @@ void UpdateSelectionToolbarPosition(MainWindow* win);
 void HideSelectionToolbar(MainWindow* win);
 void RefreshSelectionToolbarIcons(MainWindow* win);
 void DeleteSelectionToolbar(MainWindow* win);
+TempStr SelectionToolbarLayoutDumpTemp();